requirements.txt
PyYAML
six
flask
flask-sqlalchemy
gunicorn
gevent
